bias/requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
openai>=1.0.0
